requirements.txt
numpy
scipy
opencv-python
tensorflow-gpu==1.2.0
